Scraping and Storing
Worked on the initial bot.js setup and added a function to scrape the Eurogamer website using the Cheerio API. The codes are found by searching the HTML elements and then I stored them in a SQLite database.

Next: Have the bot compare the HTML and the existing database, then alert if any changes.