Slashing Commands, Mentioning, and more
Incorporated the foundation for Discord Slash commands as well as implemented the first command 'checkcodes'.

'checkcodes' simply runs the scrape functions and returns any new codes or no codes. I reworked the functions to acknowledge when a user is prompting it and when it is a normal scan.

Example: When a user prompts the scraper, it will @ the user but when its a normal scan, it will @ the Genshin role.